+++ core/src/main/java/org/apache/iceberg/avro/BuildAvroProjection.java
@@ -107,7 +107,10 @@ class BuildAvroProjection extends AvroCustomOrderSchemaVisitor<Schema, Schema.Fi
             field.name());
         // Create a field that will be defaulted to null. We assign a unique suffix to the field
         // to make sure that even if records in the file have the field it is not projected.
-        String fieldName = AvroSchemaUtil.makeCompatibleName(field.name());
+        String origFieldName = field.name();
+        boolean isValidFieldName = AvroSchemaUtil.validAvroName(origFieldName);
+        String fieldName =
+            isValidFieldName ? origFieldName : AvroSchemaUtil.sanitize(origFieldName);
         Schema.Field newField =
             new Schema.Field(
                 fieldName + "_r" + field.fieldId(),
@@ -115,7 +118,9 @@ class BuildAvroProjection extends AvroCustomOrderSchemaVisitor<Schema, Schema.Fi
                 null,
                 JsonProperties.NULL_VALUE);
         newField.addProp(AvroSchemaUtil.FIELD_ID_PROP, field.fieldId());
-        newField.addProp(AvroSchemaUtil.ICEBERG_FIELD_NAME_PROP, field.name());
+        if (!isValidFieldName) {
+          newField.addProp(AvroSchemaUtil.ICEBERG_FIELD_NAME_PROP, origFieldName);
+        }
         updatedFields.add(newField);
         hasChange = true;
       }
